* rename Relation_* classes to Reference_*. * respective changes in docs and tests
